move timeout middleware to its own module
