+++ rust-runtime/aws-smithy-http-client/src/client.rs
@@ -1,0 +1,6 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+pub(crate) mod timeout;
